Started ground work for the lynx protocol.
Note: I set the default main method to RobotInitial.Lynx_Server.ServerTest... you can change it back in project properties.

+++ RobotInitial/Model/Data/MoveParameters.cs
@@ -32,7 +32,7 @@ namespace RobotInitial.Model {
         MoveDirection LeftDirection { get; }
         int RightPower { get; }
         int LeftPower { get; }
-        int Duration { get; }
+        float Duration { get; }
         MoveDurationUnit DurationUnit { get; }
         bool BrakeAfterMove { get; }
     }
